pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "hatchling"]
build-backend = "setuptools.build_meta"

[project]
name = "kodudo"
version = "0.1.0"
description = "Cook your data into documents using Jinja2 templates"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "GPL-3.0-or-later"}
authors = [
    {name = "Tiago Tresoldi"},
]
keywords = ["template", "jinja2", "data", "rendering", "etl"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Text Processing :: Markup",
]

dependencies = [
    "jinja2>=3.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "mypy>=1.0",
    "ruff>=0.1",
    "types-PyYAML",
    "build",
    "twine",
]

[project.urls]
Homepage = "https://github.com/plataformasindigenas/kodudo"
Documentation = "https://github.com/plataformasindigenas/kodudo#readme"
Repository = "https://github.com/plataformasindigenas/kodudo"

[project.scripts]
kodudo = "kodudo.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
